+++ database/src/main/java/com/backstage/app/database/configuration/ddl/DDLConfiguration.java
@@ -24,6 +24,7 @@ import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
 
 import java.util.Map;
 
@@ -32,6 +33,7 @@ import java.util.Map;
 @RequiredArgsConstructor
 @AutoConfigureAfter(AppConfiguration.class)
 @EnableConfigurationProperties(DDLProperties.class)
+@PropertySource("classpath:flyway.properties")
 public class DDLConfiguration
 {
 	/**