+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 ### Database
 - При инициализации EntityManager по умолчанию включаются сущности из appProperties.basePackages.
 - Для @SequenceGenerator добавлена поддержка SpEL и placeholder.
+- При подключении модуля параметр spring.flyway.enabled устанавливается в false.
 
 ### Dicts
 - Метод dictDataService.getByIds при передаче пустого списка идентификаторов теперь возвращает пустой список вместо исключения.