+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - При инициализации EntityManager по умолчанию включаются сущности из appProperties.basePackages.
 - Для @SequenceGenerator добавлена поддержка SpEL и placeholder.
 - При подключении модуля параметр spring.flyway.enabled устанавливается в false.
+- Удалили параметр app.jpa.postgis-enabled, расширения JPA для PostGIS активируются автоматически.
 
 ### Dicts
 - Метод dictDataService.getByIds при передаче пустого списка идентификаторов теперь возвращает пустой список вместо исключения.